Ignore message_changed events
When a message is edited, slack resends the message setting event type
to message_changed. Since commands may not be idempotent, there is risk
in reprocessing these messages. For now, ignore message changed events.
If a plugin needs to reprocess these messages, we can decide to handle
these events at that time.